Refactor `httpie.input` into submodules
