+++ httpie/input/exceptions.py
@@ -1,0 +1,2 @@
+class ParseError(Exception):
+    pass
