+++ tests/test_auth_plugins.py
@@ -1,6 +1,6 @@
 from mock import mock
 
-from httpie.input import SEP_CREDENTIALS
+from httpie.input.constants import SEP_CREDENTIALS
 from httpie.plugins import AuthPlugin, plugin_manager
 from utils import http, HTTP_OK
 
@@ -103,7 +103,7 @@ def test_auth_plugin_require_auth_false_and_auth_provided(httpbin):
         plugin_manager.unregister(Plugin)
 
 
-@mock.patch('httpie.input.AuthCredentials._getpass',
+@mock.patch('httpie.input.argtypes.AuthCredentials._getpass',
             new=lambda self, prompt: 'UNEXPECTED_PROMPT_RESPONSE')
 def test_auth_plugin_prompt_password_false(httpbin):
