+++ tests/test_exit_status.py
@@ -5,7 +5,7 @@ from utils import TestEnvironment, http, HTTP_OK
 
 
 def test_keyboard_interrupt_during_arg_parsing_exit_status(httpbin):
-    with mock.patch('httpie.cli.parser.parse_args',
+    with mock.patch('httpie.input.cli.parser.parse_args',
                     side_effect=KeyboardInterrupt()):
         r = http('GET', httpbin.url + '/get', error_exit_ok=True)
         assert r.exit_status == ExitStatus.ERROR_CTRL_C
